feat: one function per file
