+++ apps/evm/src/libs/errors/handleContractError/constants.ts
@@ -1,0 +1,1 @@
+export const SELECTOR_LENGTH = 10;
